use of long double for karma output

+++ progs/cicada_alignment_hmm.cpp
@@ -1935,9 +1935,9 @@ struct PosteriorReducer : public PosteriorMapReduce
     }
   }
 
-  struct real_precision : boost::spirit::karma::real_policies<double>
+  struct real_precision : boost::spirit::karma::real_policies<long double>
   {
-    static unsigned int precision(double) 
+    static unsigned int precision(long double) 
     { 
       return 20;
     }
@@ -1953,7 +1953,7 @@ struct PosteriorReducer : public PosteriorMapReduce
     
     const matrix_type& matrix = posterior.matrix;
 
-    karma::real_generator<double, real_precision> real;
+    karma::real_generator<long double, real_precision> real;
     iterator_type iter(os);
 
     if (! matrix.empty()) {
